Ignore previous Comments anout bug code and issue. Its written mistakenly
TB004	Remove mandatory from Guarantor, No. of Co-Borrowers, Tenure, No. of Installments, Installment start date and End Date Fields. In the add liabilities

TB006	While entering details in the liability...when we go to the “Guarantor” tab it gives the option to select the names only of the existing customer’s…..however a Guarantor could be anybody and it does not give the option to add anyone else’s name.Remove guarantor Drop down list, Give a field to be filled manually for guarantor.